JNI Signiture "J" 에 해당하는 jlong 타입이 jni.h 에 long long 으로 선언되어 있으므로 long 타입이 아닌 long long 타입을 기본으로 보는 것이 좋을 듯.
git-svn-id: svn://35.216.44.119/DVM/dragonvillage/src/frameworks@87845 5003bd00-c432-460b-831f-0659da9acde2

+++ cocos2d-x/cocos/audio/android/jni/cddandroidAndroidJavaEngine.cpp
@@ -328,7 +328,7 @@ void AndroidJavaEngine::unloadEffect(const char* filePath)
 }
 
 void AndroidJavaEngine::playVibrate(long millisecond) {
-    JniHelper::callStaticVoidMethod(helperClassName, "playVibrate", millisecond);
+    JniHelper::callStaticVoidMethod(helperClassName, "playVibrate", (long long)millisecond);
 }
 
 void AndroidJavaEngine::cancelVibrate() {
